KPI Visualization using Excel
I visualized key performance indicators for a fictitious McDonald's company using excel, I also linked each sheet with the series point in the dashboard sheet, so if you click on any icon on the far left of the dashboard sheet, you will be taken to sheet that was linked to it. Thanks!